requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
coverage==5.5
dj-database-url==0.5.0
Django==3.2.6
django-ckeditor==6.1.0
django-email-verification==0.1.0
django-js-asset==1.2.2
django-lockout==0.1.1
Django-Verify-Email==1.0.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
factory-boy==3.2.0
Faker==8.12.1
gunicorn==20.1.0
idna==3.2
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.6
whitenoise==5.3.0
